Dual-boot: BIOS attribute for choosing OS
This commit adds interface to choose the OS.
This property reflects the user settings and not a status.
It is possible that the user can change the setting
at any point of time. But the property will be applied
only when the host boots the next time.

The interface is contained within com.ibm.Host namespace.

Signed-off-by: Pavithra Barithaya <pbaritha@in.ibm.com>
Signed-off-by: Deepak Kodihalli <dkodihal@in.ibm.com>