Add Self Boot Engine(SBE) dump interface
Self Boot Engine(SBE) is a microcontroller that sits inside the
processor to initialize it to start the booting and also acts as a
secure channel for accessing certain control functions on the processor.
During the booting or other hardware access operations SBE can encounter
errors and become unresponsive. In such situations, the debug data needs
to be collected from such SBEs to find out the root cause of the error.
This data includes hardware state, configuration, memory, etc.
The collected data is then packaged into the OpenPOWER dump format
and which is called as SBE dump.
Implement this interface to add support for managing the SBE dump.

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>
Change-Id: Ibf625c0205096a4f94bf159125cf8fdae093449c