Add Error to Software Version Interface
This commit adds a InvalidSignature error to the
Software Version interface. It will ensure that
if the error is hit during the verification of
an image, a corresponding error is available to the
user. This will replace the InternalError that the
Update Access Key function currently creates if it
fails.