Add Error to Software Version Interface (#43)
This commit adds a InvalidSignature error to the
Software Version interface. It will ensure that
if the error is hit during the verification of
an image, a corresponding error is available to the
user. This will replace the InternalError that the
signature verification currently creates if it
fails.

This commit will also add a ExpiredAccessKey error
to the Software Version interface. It will provide the
user not only with a more specific error, but metadata
indicating the direct cause. This will also replace an
InternalError that the Update Access Key verifier creates
when the criteria is not met.