Add StaticNTPServers propetry to EthernetInterface (#47)
Currently there is no way to differentiate between the Static
and DHCP assigned NTP servers at dbus

This commit adds StaticNTPServers peroperty to hold the static
configuration of NTP servers. This is a read-write property.
The existing NTPServer propery will hold both static and DHCP
assigned NTP servers. This is a read-only property.

Signed-off-by: Sunitha Harish <sunharis@in.ibm.com>
Change-Id: If6ac588c268473802ebb1b8f08144df594ac943a

Co-authored-by: Sunitha Harish <sunharis@in.ibm.com>