Power mode state lock support (#61)
* Power mode state lock support

The power mode state can be locked such that it cannot be changed. Two
methods are provided. One method is for locking the mode and another for
querying the current state of the lock.

Signed-off-by: Ben Tyner <ben.tyner@ibm.com>

* Support for PowerMode property NotAllowed error

Add a NotAllowed error to the PowerMode property. Currently this
error is used to indicate to the GUI that the power mode is locked.

Signed-off-by: Ben Tyner <ben.tyner@ibm.com>