June 17: 1020 rebase (#62)
* Power mode state lock support (#61)

* Power mode state lock support

The power mode state can be locked such that it cannot be changed. Two
methods are provided. One method is for locking the mode and another for
querying the current state of the lock.

Signed-off-by: Ben Tyner <ben.tyner@ibm.com>

* Support for PowerMode property NotAllowed error

Add a NotAllowed error to the PowerMode property. Currently this
error is used to indicate to the GUI that the power mode is locked.

Signed-off-by: Ben Tyner <ben.tyner@ibm.com>

* Add additional DHCP configuration options (#60)

The existing D-Bus DHCPConf enumearation values do not allow IPv6
Stateless DHCP to be assigned. The IPv4 entries are sufficient, as
DHCP for the v4 stack is truly a boolean.

IPv6 is, at minimum, a ternary control. Adding new enumerations allows
the IPv6 configruation to be done in a more abstract fashion than the
current IPv6AcceptRA D-Bus entry. This D-Bus entry is a direct
reflection of how systemd-networkd manages Routint Advertisement, and
by extension DHCPv6. Only phosphor-network needs this knowledge.

Tested:
The changes compile.
Also, in combination with changes made to phosphor-network, many
busctl set/get property commands were issued to confirm the
enumerations were performing as expected.

Change-Id: I42a46549fed8b2d2187ea6a7eb7278c63b84bbef
Signed-off-by: Johnathan Mantey <johnathanx.mantey@intel.com>

Co-authored-by: Johnathan Mantey <johnathanx.mantey@intel.com>

Co-authored-by: Ben Tyner <ben.tyner@ibm.com>
Co-authored-by: Sunitha Harish <32155533+sunharis@users.noreply.github.com>
Co-authored-by: Johnathan Mantey <johnathanx.mantey@intel.com>