PCIeLink and PCIeTopology DBus Interfaces
This commit
- adds support for com.ibm.Control.Host.PcieLink
interface that hosts a new boolean property called linkReset.
- defines PCIeTopology DBus interface.
This interface is used by applications on bmc to
request the pcie topology information.
- defines new enum property called CableStatus under the Cable Interface.
This also adds the downstream chassis_associations and
connector associations.
- changes the type and default value for property
LanesInUse in order to accommodate the situation where reason
of not getting this value is unknown.
- defines property to hold PCIe link status value
for a given PCIe slot.
- adds property BusId to PCIeSlot interfce, the
property will be required to detect the bus on which the
slot resides.

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>