software: Add Minimum Version interface (#114)
The minimum version is an optional meson option (bmc-msl)[1] that
prevents firmware updates to a level lower than the one set via this
option.
The error logged is a Software Incompatible[2] error.

Add the minimum version as a property of a new interface. This will be
populated by the code updater app. It can then be mapped to the
SoftwareInventory LowestSupportedVersion Redfish property and be
available to the user via Redfish and the GUI so that the user can know
in advance which version of software is not supported on their system
and avoid having to debug a code update failure after they attempt to
install an incompatible version.

```
1: https://github.com/openbmc/phosphor-bmc-code-mgmt/blob/85c71a13e0938cc4d36caf6b8e735e9740b2e351/meson.options#L100
2: https://github.com/openbmc/phosphor-dbus-interfaces/blob/1c140b9766a15d1cbb8546fa02d5050d772a171d/yaml/xyz/openbmc_project/Software/Version.errors.yaml#L1
```

Change-Id: I1f6aacf8ad7fa63e58d832409698f4d923184ec2

Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>