com.ibm: Add MemoryBufferSBE dump type (#127)
The Self Boot Engine (SBE) is contained on certain memory
buffers. When the SBE becomes unresponsive due to hardware
or software issues, there's a need to collect diagnostic
dumps from it. To address this need and differentiate such
dumps from others, this commit adds the `MemoryBufferSBE`
type to the com.ibm.Dump.Create interface. The structure
and input to collect this dump is different from the
processor SBE dump so making it a new dump type.

Change-Id: Ib07a0caa1521c5ae5f6e0682029b5c9e497152e8

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>