State: Add RBMC FailoversPaused property (#162)
When redundancy is enabled, there may be periods when either failovers
are not allowed, such as in the middle of a code update, or won't work
because the passive BMC is temporarily offline, such as when the passive
BMC reboots. A timer would be put on how long redundancy could still be
considered enabled in this latter case in case the passive BMC never
comes back.  Redundancy is left enabled initially so as to not trigger
any intervention that could be necessary when redundancy is lost just
due to a BMC reboot.  Any time the passive BMC goes offline a full file
sync would be necessary when it comes back.

Change-Id: I713cefad67298e62a03596654a0629cc60fb4caa

Signed-off-by: Matt Spinler <spinler@us.ibm.com>