State:Add FailoversAllowed to RBMC interfaces
This property indicates if a failover is allowed at any instance in
time.  If redundancy is disabled, it will of course be false.

To stage in the conversion of the FailoversPaused to FailoversAllowed,
leave in the original property while adding the new one.  The old one
will be removed when everything points to the new one.

Note that the Sibling RBMC interface will eventually be removed in place
of using separate interfaces.

Change-Id: I19ed39e8ae22a1902e33b4f7ca1a06b66181f139
Signed-off-by: Matt Spinler <spinler@us.ibm.com>